requirements.txt
Django == 3.1

wheel == 0.35.1
coreapi == 2.3.3
Markdown ==  3.2.2
Pygments == 2.6.1
django-filter == 2.3.0
djangorestframework == 3.11.1
drf-yasg == 1.17.1

PyJWT~=1.7.1

djangorestframework-simplejwt ==4.4.0
